refactor(web): move useTags in with the filters that offer them
`/api/query/valid-tags` returns the tags the tag filter lists, so the hook
fetching it belongs with the rest of the filter state rather than in
web/src/hooks, which the frontend standards keep for hooks with no feature
home.

Now a named export, matching its neighbours; both call sites already
destructured it.
